Refactor: AttendanceStatusResourceをStatusResourceに置き換え、関連するAPIを更新
